sprint-5-pb-aws-furg-ifrs-uffs/equipe-1/api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.101.0
h11==0.14.0
idna==3.4
pydantic==2.1.1
pydantic-core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
python-dotenv==1.0.0
sagemaker==2.178.0
boto3==1.28.30
numpy==1.25.0
pandas==2.0.3